Implement the perturbed_A matrix for the modified_alternate_phase_projection
In this commit, we use a matrix called perturbed_A to simulate an unknown matrix
A that is an estimate of the real A. As of now, the error is randomized and divided
by 10^4. This shows that the error gets lower as we increase masks; however, if we
use a smaller error by default we can get consistent recovery with lower masks too.