Create new recovery demo tab
This tab shows the user the various graphs, errors, and results from running
the selected algorithm. I had to change some of the functions to return the
iterated values of the reconstructed signal to show the error change over time.
Currently it appears the mask recovery part is broken and needs to be fixed.